open eml files in neomutt
The eml file (mime-type message/rfc822) is converted to a mailbox file
(.mbox) using formail and the mailbox file is opened in a new instance
of neomutt.

An entry was added to the local mailcap file (~/.mailcap). When invoked
this results in the mailbox file being left in the directory. (The
mailbox file name is the eml file name with a '.mbox' extension.)

A local wrapper script was added which performs the same function,
except that the mailbox file is created in a temp directory and deleted
when neomutt exits. (On the author's system Midnight Commander was
configured to use this wrapper script for email files.)