must now change to pandoc ft to create :Pandoc cmd
It was previously necessary to set filetype to 'markdown.pandoc' to
trigger vim-pandoc's 'command' module to create the :Pandoc command.
This behaviour has clearly changed recently because the :Pandoc command
disappeared. Setting the filetype to 'pandoc' is now necessary to create
the :Pandoc command, following which it is set to markdown.pandoc as
previously done to avoid unintended changes in behaviour.